Split package-io.ts into a focused module directory
The 1582-line `src/cli/utils/package-io.ts` was the single largest code
file on this branch (1.7× the next largest) and the file Phase 2's MCP
shell-out path will exercise most. This is a mechanical split into 10
files following the pre-existing `// ---` section headers, with a
re-export `index.ts` that keeps every existing caller (10 commands, 3
test files, output.ts) working unchanged via Node's directory-as-module
resolution.

New layout:

    src/cli/utils/package-io/
      errors.ts            45 lines  PackageIOError + codes
      tree.ts             127 lines  walkBlocks, findBlockById, lookups
                                     + CONTAINER_*_KEYS constants
      auto-id.ts           84 lines  nextAutoBlockId, assignMissingIds
      state-validation.ts 199 lines  validatePackageState + classifiers
      disk.ts             273 lines  readPackage, writePackage
      mutators.ts         641 lines  appendBlock/Step/Choice + edit/remove
                                     + --if-absent helpers, JSONPath fmt
      move.ts             234 lines  moveBlock + cycle-safety
      compose.ts           40 lines  mutateAndValidate
      create-state.ts      39 lines  newPackageState
      index.ts             53 lines  Re-export the public surface

Strict acyclic layering: errors → tree → auto-id → state-validation →
disk → mutators → move → compose → create-state → index. `tsc --noEmit`
catches any import-cycle regression.

Public API unchanged. Every export from the original file is re-exported
verbatim through `index.ts`. The 18 stable `PackageIOErrorCode` strings,
the function signatures, and the type names are all preserved — these
are part of the MCP-shell-out wire contract that Phase 3 will pass
through.

Resolves the modularity concern flagged in the post-audit code review.
The next-largest concern (`json-guide.schema.ts` at 927 lines, with its
hand-maintained `KNOWN_FIELDS` mirror) is intentionally deferred.

Verified:
- `npm run typecheck` clean
- `npm run lint` clean
- `npm run prettier-test` clean
- `npm run test:ci` green: 136 suites / 2671 passed, 0 failures
- `npm run build:cli` clean; `--version` returns 1.1.0
- Smoke against `/tmp/cli-split-smoke/p`: requirement vocabulary check,
  leaf --if-absent retry no-op, quiz two-correct rejection, and
  validate output collapse all behave identically to pre-split.
- All 10 bundled-interactives still validate.
